refactor: move toJSON method to operationResponseMapper.ts
