+++ packages/specs/schema/.barrelsby.json
@@ -1,0 +1,5 @@
+{
+  "directory": ["./src", "./src/components"],
+  "exclude": ["__mock__", "__mocks__", ".spec.ts", "test", ".benchmark.ts"],
+  "delete": true
+}
