+++ packages/specs/schema/package.json
@@ -19,7 +19,7 @@
     "clean": "rm -rf lib",
     "build": "yarn clean && yarn barrels && yarn build:ts && yarn run build:browser",
     "build:browser": "webpack",
-    "barrels": "yarn barrelsby --delete -d ./src -e \"\\.spec\\.ts\" -e\"test\" -e \"__mock__\" -e \".benchmark.ts\"",
+    "barrels": "yarn barrelsby --config .barrelsby.json",
     "test": "cross-env NODE_ENV=test yarn jest --max-workers=2 && jest-coverage-thresholds-bumper",
     "build:ts": "tsc --build tsconfig.json && tsc --build tsconfig.esm.json",
     "lint": "eslint '**/*.{ts,js}'",
