+++ packages/specs/schema/src/components/index.ts
@@ -1,0 +1,19 @@
+/**
+ * @file Automatically generated by barrelsby.
+ */
+
+export * from "./anyMapper";
+export * from "./classMapper";
+export * from "./genericsMapper";
+export * from "./inheritedClassMapper";
+export * from "./itemMapper";
+export * from "./lazyRefMapper";
+export * from "./mapMapper";
+export * from "./objectMapper";
+export * from "./ofMapper";
+export * from "./operationInFilesMapper";
+export * from "./operationInQueryMapper";
+export * from "./operationParameterMapper";
+export * from "./operationResponseMapper";
+export * from "./propertiesMapper";
+export * from "./schemaMapper";
