+++ packages/specs/schema/src/index.ts
@@ -11,7 +11,10 @@ export * from "./components/lazyRefMapper";
 export * from "./components/mapMapper";
 export * from "./components/objectMapper";
 export * from "./components/ofMapper";
-export * from "./components/parameterMapper";
+export * from "./components/operationInFilesMapper";
+export * from "./components/operationInQueryMapper";
+export * from "./components/operationParameterMapper";
+export * from "./components/operationResponseMapper";
 export * from "./components/propertiesMapper";
 export * from "./components/schemaMapper";
 export * from "./constants/httpMethods";
