+++ packages/specs/schema/src/utils/getJsonSchema.ts
@@ -1,16 +1,5 @@
 import {Type} from "@tsed/core";
-import "../components/anyMapper";
-import "../components/classMapper";
-import "../components/genericsMapper";
-import "../components/inheritedClassMapper";
-import "../components/itemMapper";
-import "../components/lazyRefMapper";
-import "../components/mapMapper";
-import "../components/objectMapper";
-import "../components/ofMapper";
-import "../components/parameterMapper";
-import "../components/propertiesMapper";
-import "../components/schemaMapper";
+import "../components";
 import type {JsonEntityStore} from "../domain/JsonEntityStore";
 import {SpecTypes} from "../domain/SpecTypes";
 import {JsonSchemaOptions} from "../interfaces/JsonSchemaOptions";
